all: introduce "nm-default-0.h" as replacement for "nm-default.h"
Almost all our sources should as first include "nm-default.h".
That drags in "config.h" and other headers that should always
be available.

However, as we build different parts, "nm-default.h" gets controlled
by the NETWORKMANAGER_COMPILATION define which autotools/meson needs
to specify in the build options.

That is confusing.

One advantage of that was, that theoretically the same sources can
be built twice, with different behavior. However, we should avoid doing
that altogether and build static libraries (once) that we link multiple
times.

Instead, introduce different "nm-default-*.h" headers that don't require
special defines and behave in a consistent way.